docs: enable use with Podman Compose and explain DB_HOST (#58)
This PR is simple and does two things:

* mentions how DB_HOST behaves to access the database
* mentions what is needed to run Wiki.js on Podman

The first issue gave me cryptic errors about the wiki container being unable to access the database container, as it isn't quite obvious. The second was due to my distribution not installing aardvark by default with podman, as it is a relatively new DNS backend for podman.